src/utils/basic_utils.py
# Copyright (C) 2025-present by FringeLabs@Github, < https://github.com/FringeLabs >.
#
# This file is part of < https://github.com/FringeLabs/FLASK-PRP > project,
# and is released under the "GNU v3.0 License Agreement".
# Please see < https://github.com/FringeLabs/FLASK-PRP/blob/main/LICENSE >
#
# All rights reserved.

import yaml


def load_subjects(yaml_path: str) -> dict:
    with open(yaml_path, encoding="utf-8") as fr:
        return yaml.safe_load(fr)
