requirements.txt
pyyaml
pypdf2
flask
flask-socketIO
python-dotenv
passlib
gunicorn
openpyxl
pymongo